Pick the chunk index width adaptively: int32 normally, int64 past 2^31
The chunk-local build indexed its flat point load with int, so a chunk
past INT_MAX points hit a hard FatalError (and, before the guard,
signed-overflow UB). The re-chunk pass keeps real chunks far under the
budget, but a coincident-coordinate cluster can't be split by position
at any layer, so the ceiling was reachable by degenerate input.

Template ChunkBuilder on the point-index type and dispatch per chunk
from its planned point count: chunks at or under MaxNarrowChunkPoints
(2^31-1) keep int32 indices — identical codegen and RAM to before —
and a wider chunk gets the int64 instantiation, proceeding bounded by
RAM alone exactly like prior untwine's 64-bit PointId BU. The fatal
guard is gone; nothing about a big chunk is an error anymore. The
sampling grid becomes a local map since VoxelInfo::Grid maps to int.

bu::FileInfo's numPoints/start widen to uint64_t so the mmap length
and the flat index space are sized correctly past 2^31 (they were int
too, which would have overflowed before ChunkBuilder ever ran).

The wide path was validated by forcing MaxNarrowChunkPoints to 100
locally and re-running the lossless check: all 21 dims exact.

+++ bu/Processor.cpp
@@ -338,7 +338,7 @@ Processor::writeOctantCompressed(const OctantInfo& o, Index& index, IndexIter po
         // to the records buffer.  Otherwise, advance the position.
         while (true)
         {
-            if (pos == index.end() || *pos >= fii->start() + fii->numPoints())
+            if (pos == index.end() || (uint64_t)*pos >= fii->start() + fii->numPoints())
             {
                 count += std::distance(begin, pos);
                 appendCompressed(records, *fii, begin, pos);
@@ -353,7 +353,7 @@ Processor::writeOctantCompressed(const OctantInfo& o, Index& index, IndexIter po
                     fii++;
                     if (fii == fiiEnd)
                         goto flush;
-                } while (*begin >= fii->start() + fii->numPoints());
+                } while ((uint64_t)*begin >= fii->start() + fii->numPoints());
             }
             pos++;
         }
